refactor[#148] : MarkingFormCloseModal 분리
